Removed menu module

+++ src/migrations/m180806_105831_remove_language_tables.php
@@ -12,7 +12,10 @@ class m180806_105831_remove_language_tables extends Migration
 
         $this->dropTable('{{blog_post_lang}}');
         $this->dropTable('{{blog_category_lang}}');
-        $this->dropTable('{{menu_lang}}');
+
+        if ($this->getDb()->getTableSchema('{{menu_lang}}')) {
+            $this->dropTable('{{menu_lang}}');
+        }
 
         if ($this->getDb()->getTableSchema('{{new_lang}}')) {
             $this->dropTable('{{new_lang}}');
